requirements.txt
pygame
psycopg2-binary
numpy
matplotlib
python-dotenv